adds jsonapi-rb to output jsonapi formatted data from rails
REST serializer on the ember side couldn’t seem to get happy about
relationships unless the incoming data was in JSONAPI format so we’re
switching back to that again and back to JSONAPI serializer on the
ember side. Relationships seem to be working now and users can fully
login and out.